Pull reactions logic into separate file
