+++ src/spellbot/__init__.py
@@ -46,6 +46,7 @@ from spellbot.constants import (
     THUMB_URL,
 )
 from spellbot.data import Channel, Data, Event, Game, Server, Tag, User
+from spellbot.reactions import safe_clear_reactions, safe_remove_reaction
 
 load_dotenv()
 
@@ -349,27 +350,6 @@ class SpellBot(discord.Client):
         except (discord.errors.NotFound, discord.errors.HTTPException) as e:
             logger.exception("warning: discord: could fetch user", e)
             return None
-
-    async def safe_remove_reaction(
-        self, message: discord.Message, emoji: str, user: discord.User
-    ) -> None:  # pragma: no cover
-        try:
-            await message.remove_reaction(emoji, user)
-        except (
-            discord.errors.HTTPException,
-            discord.errors.Forbidden,
-            discord.errors.NotFound,
-            discord.errors.InvalidArgument,
-        ) as e:
-            logger.exception("warning: discord: could not remove reaction", e)
-
-    async def safe_clear_reactions(
-        self, message: discord.Message
-    ) -> None:  # pragma: no cover
-        try:
-            await message.clear_reactions()
-        except (discord.errors.HTTPException, discord.errors.Forbidden) as e:
-            logger.exception("warning: discord: could not clear reactions", e)
 
     async def safe_edit_message(
         self, message: discord.Message, *, reason: str = None, **options
@@ -690,7 +670,7 @@ class SpellBot(discord.Client):
             now = datetime.utcnow()
             expires_at = now + timedelta(minutes=server.expire)
 
-            await self.safe_remove_reaction(message, emoji, author)
+            await safe_remove_reaction(message, emoji, author)
 
             if emoji == "➕":
                 if any(user.xid == game_user.xid for game_user in game.users):
@@ -742,7 +722,7 @@ class SpellBot(discord.Client):
                 for discord_user in found_discord_users:
                     await discord_user.send(embed=game.to_embed())
                 await self.safe_edit_message(message, embed=game.to_embed())
-                await self.safe_clear_reactions(message)
+                await safe_clear_reactions(message)
             else:
                 session.commit()
                 await self.safe_edit_message(message, embed=game.to_embed())
@@ -1104,7 +1084,7 @@ class SpellBot(discord.Client):
                     for discord_user in found_discord_users:
                         await discord_user.send(embed=game.to_embed())
                     await self.safe_edit_message(post, embed=game.to_embed())
-                    await self.safe_clear_reactions(post)
+                    await safe_clear_reactions(post)
                 else:
                     session.commit()
                     await self.safe_edit_message(post, embed=game.to_embed())
